api/requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.2.0
channels[daphne]==4.0.0
hume==0.4.1
matplotlib==3.8.2
pillow==10.2.0